+++ src/Generics/SOP/NP.hs
@@ -74,6 +74,8 @@ import qualified Data.Vector as V
 import GHC.Exts (Any)
 import Unsafe.Coerce
 
+import Control.DeepSeq (NFData(..))
+
 import Generics.SOP.BasicFunctors
 import Generics.SOP.Classes
 import Generics.SOP.Constraint
@@ -151,6 +153,11 @@ instance (All (Eq `Compose` f) xs, All (Ord `Compose` f) xs) => Ord (NP f xs) wh
   compare xs ys =
     mconcat (hcollapse (hczipWith (Proxy :: Proxy (Ord `Compose` f)) (\ x y -> K (compare x y)) xs ys))
 
+-- | @since 0.2.5.0
+instance All (NFData `Compose` f) xs => NFData (NP f xs) where
+    rnf Nil       = ()
+    rnf (x :* xs) = rnf x `seq` rnf xs
+
 -- | A product of products.
 --
 -- This is a 'newtype' for an 'NP' of an 'NP'. The elements of the
@@ -170,6 +177,10 @@ newtype POP (f :: (k -> *)) (xss :: [[k]]) = POP (NP (NP f) xss)
 deriving instance (Show (NP (NP f) xss)) => Show (POP f xss)
 deriving instance (Eq   (NP (NP f) xss)) => Eq   (POP f xss)
 deriving instance (Ord  (NP (NP f) xss)) => Ord  (POP f xss)
+
+-- | @since 0.2.5.0
+instance (NFData (NP (NP f) xss)) => NFData (POP f xss) where
+    rnf (POP xss) = rnf xss
 
 -- | Unwrap a product of products.
 unPOP :: POP f xss -> NP (NP f) xss